Module oak_dice::cert

Constants and helper functions to work with CWT-based DICE certificates.

Structs

KeyUsage

Intended usage of a key.

Constants

ACPI_MEASUREMENT_ID

The CWT private claim ID for the concatenated hash of the commands for building the ACPI tables.

APPLICATION_KEY_ID

The CWT private claim ID of the application keys.

CONTAINER_IMAGE_LAYER_ID

The CWT private claim ID for the container image.

ENCLAVE_APPLICATION_LAYER_ID

The CWT private claim ID for an enclave application layer.

EVENT_ID

The CWT private claim ID of the Event digest.

FINAL_LAYER_CONFIG_MEASUREMENT_ID

The CWT private claim ID for the measurement of the application or container config.

INITRD_MEASUREMENT_ID

The CWT private claim ID for the initial RAM file system measurement.

KERNEL_COMMANDLINE_ID

The CWT private claim ID for the raw kernel command-line.

KERNEL_COMMANDLINE_MEASUREMENT_ID

The CWT private claim ID for the kernel command-line measurement.

KERNEL_LAYER_ID

The CWT private claim ID for the Kernel layer.

KERNEL_MEASUREMENT_ID

The CWT private claim ID for the kernel measurement.

KEY_ID_LENGTH

Length of the unique ID for ECDSA keys generated.

KEY_USAGE_ID

ID for the bitstring used to describe the intended usage for the key represented by the certificate.

LAYER_2_CODE_MEASUREMENT_ID

The CWT private claim ID for the measurement of the layer 2 binary.

LAYER_3_CODE_MEASUREMENT_ID

The CWT private claim ID for the measurement of the layer 3 binary.

MEMORY_MAP_MEASUREMENT_ID

The CWT private claim ID for the physical memory map (e820 table).

SETUP_DATA_MEASUREMENT_ID

The CWT private claim ID for the kernel setup data measurement.

SHA2_256_ID

The CWT private claim ID for SHA2_256 digests.

SUBJECT_PUBLIC_KEY_ID

ID for the CWT private claim corresponding to the Subject of the CWT.

SYSTEM_IMAGE_LAYER_ID

The CWT private claim ID for the Oak Containers system image.

Functions

cose_key_to_hpke_public_key

Converts a COSE_Key to a serialized HPKE KEM public key.

cose_key_to_verifying_key

Converts a COSE_Key to a ECDSA verifying key.

derive_kem_public_key_id

Derives an ID from an HPKE KEM public key.

derive_verifying_key_id

Derives an ID from a verifying key.

generate_ecdsa_key_pair

Generates private/public ECDSA key pair.

generate_kem_certificate

Generates a CWT certificate representing a Key Encapsulation Mechanism (KEM) for Hybrid Public-Key Encryption (HPKE) from a serialized public key.

generate_signing_certificate

Generates a CWT certificate representing an ECDSA signing key, such as an embedded certificate authority (ECA), and returns the certificate and its associated private key.

get_claims_set_from_certificate_bytes

Parses a bytes slice as a CWT certificate and extracts the payload as a set of claims.

get_public_key_from_claims_set

Extracts the certified public key from the claims set of a certificate.

hpke_public_key_to_cose_key

Converts a serialized HPKE KEM public key to a COSE_Key representation

verifying_key_to_cose_key

Converts an ECDSA verifying key to a COSE_Key representation.