Function oak_dice::cert::generate_signing_certificate

pub fn generate_signing_certificate(
    issuer_eca_key: &SigningKey,
    issuer_id: String,
    verifying_key: &VerifyingKey,
    additional_claims: Vec<(ClaimName, Value)>,
) -> Result<CoseSign1, CoseError>

Generates a CWT certificate representing an ECDSA signing key, such as an embedded certificate authority (ECA), and returns the certificate and its associated private key.

Arguments

issuer_eca_key - The private key that the issuer uses to sign issued certificates.

issuer_id - A string identifying the key used by the issuer. This would typically be the hex encoding of the output from calling derive_verifying_key_id using the issuer’s public key.

additional_claims - Any additional claims that must be included in the certificate. This is typically used to provide measurements of the components in the layer associated with th certificate.