Module oak_dice::evidence

C-like structs for representing DICE data in environments where we don’t have protocol buffer support.

Structs

ApplicationKeys

Certificates for the ECDSA keys that can be used for an application for signing or encryption.

ApplicationPrivateKeys

ECDSA private keys that can be used for an application for signing or encryption.

CertificateAuthority

Private key that can be used by a layer to sign a certificate for the next layer.

CompoundDeviceIdentifier

Evidence

Wrapper for passing the attestation evidence from the Restricted Kernel to the application.

LayerEvidence

Attestation evidence for non-root layers.

RestrictedKernelDiceData

Wrapper for passing the attestation evidence and private keys from the Restricted Kernel to the application.

RootLayerEvidence

Attestation evidence generated by Stage 0.

Stage0DiceData

Wrapper for passing DICE info from Stage0 to the next layer (Stage 1 or the Restricted Kernel).

Enums

TeePlatform

The hardware platform providing the Trusted Execution Environment.

Constants

CDI_SIZE

The size for compound device identifiers.

CERTIFICATE_SIZE

The maximum size of a serialized CWT certificate.

DICE_DATA_CMDLINE_PARAM

The name of the kernel command-line parameter that is used to send the physical address of the Stage 0 DICE data struct.

DICE_DATA_LENGTH_CMDLINE_PARAM

The name of the kernel command-line parameter containing dice data length.

EVENTLOG_CMDLINE_PARAM

physical address of the Stage 0 event log.

P256_PRIVATE_KEY_SIZE

The actual size used when encoding a Nist P256 private key.

PRIVATE_KEY_SIZE

The maximum size of an ECDSA private key.

PUBLIC_KEY_SIZE

The maximum size of a serialized COSE Key object representing an ECDSA public key.

REPORT_SIZE

The maximum size of the signed attestation report.

STAGE0_DICE_PROTO_MAGIC

The magic number used to identify the Stage 0 DICE proto in memory.

STAGE0_MAGIC

The magic number used to identify the Stage 0 DICE data in memory.

X25519_PRIVATE_KEY_SIZE

The actual size used when encoding a X25519 private key.