Struct oak_dice::evidence::Evidence

#[repr(C)]
pub struct Evidence {
    pub root_layer_evidence: RootLayerEvidence,
    pub restricted_kernel_evidence: LayerEvidence,
    pub application_keys: ApplicationKeys,
}

Wrapper for passing the attestation evidence from the Restricted Kernel to the application.

Fields
root_layer_evidence: RootLayerEvidence

Evidence about Stage 0 and the initial state of the VM.

restricted_kernel_evidence: LayerEvidence

The evidence about the Restricted Kernel.

application_keys: ApplicationKeys

Keys (and associated certificates) that can be used by the application for encryption or signing.