Struct oak_dice::evidence::RootLayerEvidence

#[repr(C)]
pub struct RootLayerEvidence {
    pub tee_platform: u64,
    pub remote_attestation_report: [u8; 2048],
    pub eca_public_key: [u8; 256],
    /* private fields */
}

Attestation evidence generated by Stage 0.

Fields
tee_platform: u64

The hardware platform that generated the remote attestation report.

remote_attestation_report: [u8; 2048]

Byte representation of the signed attestation report provided by the TEE platform.

eca_public_key: [u8; 256]

Serialized COSE Key representation of the ECDSA public key that can be used to verify the signature of the next layer’s certificate.

The SHA2-256 digest of this field must included as additional data in the signed remote attestation report to bind the key to the report. This means that the attestation report effectively acts as a non-standard certificate for this key.

Implementations

impl RootLayerEvidence

pub fn get_tee_platform(&self) -> Option<TeePlatform>